+++ src/main/java/com/manhduc205/ezgear/dtos/PurchaseOrderItemDTO.java
@@ -18,5 +18,5 @@ public class PurchaseOrderItemDTO {
 
     private Integer quantity;
 
-    private BigDecimal unitPrice;
+    private Long  unitPrice;
 }
